Added `JtComponent` class
This class is used to create a new web component. Its constructor takes a URL as an argument and it does the basic component setup for the user. It also exposes certain variables and functions for convenience.

I also updated all the current custom components to use the new `JtComponent` class